feat: single dispatch engines
