+++ src/sdk/harvest/__init__.py
@@ -1,7 +1,8 @@
 from pydantic.types import *  # type: ignore
 from .media import *  # type: ignore
+from .model import Meta
 from .fields import CIDString
-from .types import Collector, Meta, Collection
+from .types import Collector,Collection
 from .collectors import merge, map, load
