+++ src/sdk/processing/types.py
@@ -3,14 +3,14 @@ from src.core.types import Protocol, Any, Path
 
 
 class Engine(Protocol, metaclass=ABCMeta):
-    """Abstract middleware class class to handle standard actions for media processing.
+    """Abstract middleware class to handle standard actions for media processing.
 
     Any type of input, for example, video, image, music, etc., needs some type of processing to be transmitted or consumed.
     We could define any steps or logic needed to process our media.
     """
 
     @abstractmethod
-    def __init__(self, media: Path):
+    def __init__(self, path: Path):
         """Initialize engine with any media path."""
         ...
 
@@ -20,13 +20,19 @@ class Engine(Protocol, metaclass=ABCMeta):
         ...
 
     @abstractmethod
-    def __call__(self, *options: Any):
-        """Execute any action over media proxy using instructions
+    def __call__(self, **options: Any):
+        """Add options to the processing context
+
         eg.
-            engine = Engine(Media("image.png"))
-            engine(Instruction(method, params))
+
+        engine = Engine(Path(..))
+        with engine(max_muxing_queue_size=10) as video:
+            ...
         """
         ...
+
+    def __getattr__(self, name: str) -> Any:
+        """Delegate calls to any internal processing logic"""
 
     @abstractmethod
     def __exit__(self):
